+++ amethyst_assets/src/loader.rs
@@ -10,7 +10,6 @@ use crate::{
 
 /// The asset loader, holding the sources and a reference to the `ThreadPool`.
 pub struct Loader {
-    directory: Arc<Directory>,
     hot_reload: bool,
     pool: Arc<ThreadPool>,
     sources: FnvHashMap<String, Arc<dyn Source>>,
@@ -23,12 +22,22 @@ impl Loader {
     where
         P: Into<PathBuf>,
     {
-        Loader {
-            directory: Arc::new(Directory::new(directory)),
+        Self::with_default_source(Directory::new(directory), pool)
+    }
+
+    /// Creates a new asset loader, using the provided source
+    pub fn with_default_source<S>(source: S, pool: Arc<ThreadPool>) -> Self
+    where
+        S: Source,
+    {
+        let mut loader = Loader {
             hot_reload: true,
             pool,
             sources: Default::default(),
-        }
+        };
+
+        loader.set_default_source(source);
+        loader
     }
 
     /// Add a source to the `Loader`, given an id and the source.
@@ -39,6 +48,14 @@ impl Loader {
     {
         self.sources
             .insert(id.into(), Arc::new(source) as Arc<dyn Source>);
+    }
+
+    /// Set the default source of the `Loader`.
+    pub fn set_default_source<S>(&mut self, source: S)
+    where
+        S: Source,
+    {
+        self.add_source(String::new(), source);
     }
 
     /// If set to `true`, this `Loader` will ask formats to
@@ -129,14 +146,10 @@ impl Loader {
             handle,
         );
 
-        let source = match source {
-            "" => self.directory.clone(),
-            source => self.source(source),
-        };
-
         progress.add_assets(1);
         let tracker = progress.create_tracker();
 
+        let source = self.source(source);
         let handle_clone = handle.clone();
         let processed = storage.processed.clone();