+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 
 * Introduce `application_dir` utility ([#1213])
 * Derive `Copy`, `PartialEq`, `Eq`, `Serialize`, `Deserialize` for `Flipped` component. ([#1237])
+* A way to change the default `Source` using `set_default_source` and `with_default_source`. ([#1256])
 
 ### Changed
 
@@ -25,6 +26,8 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 
 [#1213]: https://github.com/amethyst/amethyst/pull/1213
 [#1237]: https://github.com/amethyst/amethyst/pull/1237
+[#1256]: https://github.com/amethyst/amethyst/pull/1256
+[#1267]: https://github.com/amethyst/amethyst/pull/1267
 
 ## [0.10.0] - 2018-12
 
@@ -108,7 +111,6 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#1188]: https://github.com/amethyst/amethyst/pull/1188
 [#1198]: https://github.com/amethyst/amethyst/pull/1198
 [#1224]: https://github.com/amethyst/amethyst/pull/1224
-[#1267]: https://github.com/amethyst/amethyst/pull/1267
 [winit_018]: https://github.com/tomaka/winit/blob/v0.18.0/CHANGELOG.md#version-0180-2018-11-07
 [glutin_019]: https://github.com/tomaka/glutin/blob/master/CHANGELOG.md#version-0190-2018-11-09